requirements.txt
streamlit>=1.50.0
python-dotenv>=1.1.1
groq>=0.33.0